feat(iac_scan_ci): use predefined env variables
Issue: IAC-120
Changelog: skip
