+++ changelog.d/20230718_171624_paul.beslin.ext_iac_scan_ci_command.md
@@ -1,0 +1,9 @@
+### Added
+
+- A new command is added to IaC scans: `ggshield iac scan ci`. It is meant to simplify using IaC scans in CI environments.
+  It allows the same arguments as hook subcommands (in particular, `--all` to perform a scan on the whole repository).
+  Supported CIs are: - Azure - Bitbucket - CircleCI - Drone - Github - Gitlab - Jenkins - Travis
+
+### Changed
+
+- Code is shared between iac scan CI and secret scan CI. Commit range functions are now inside `ggshield/core/git_hooks/ci.py`.
